[Refactor/#48]: 댓글 조회, 작성 통일
