+++ src/features/Feed/api/comments.api.ts
@@ -16,34 +16,42 @@ const searchMemoComments = new SearchMemoComment(baseOptions);
 const postReviewComment = new PostReviewComment(authOptions);
 const postMemoComment = new PostMemoComment(authOptions);
 
-interface GetReviewComments {
-  reviewId: string;
+interface GetComments {
+  id: string;
   page?: number;
   depth?: number;
   parentId?: string;
+  isReview: boolean;
 }
 
-interface GetMemoComments {
-  memoId: string;
-  page?: number;
-  depth?: number;
-  parentId?: string;
-}
-
-export const getReviewComments = async ({
-  reviewId,
+export const getComments = async ({
+  id,
   page,
   depth,
   parentId,
-}: GetReviewComments) => {
+  isReview,
+}: GetComments) => {
   try {
-    const result = await searchReviewComments.fetch({
-      review_id: reviewId,
-      limit: page ? LIMIT : undefined,
-      page,
-      depth,
-      parent: parentId,
-    });
+    let result;
+
+    if (isReview) {
+      result = await searchReviewComments.fetch({
+        review_id: id,
+        limit: page !== undefined ? LIMIT : undefined,
+        page,
+        depth,
+        parent: parentId,
+      });
+    } else {
+      result = await searchMemoComments.fetch({
+        memo_id: id,
+        limit: page !== undefined ? LIMIT : undefined,
+        page,
+        depth,
+        parent: parentId,
+      });
+    }
+
     return result;
   } catch (err) {
     toast.error(ERROR_ALERT);
@@ -56,38 +64,15 @@ export const getReviewComments = async ({
   }
 };
 
-export const getMemoComments = async ({
-  memoId,
-  page,
-  depth,
-  parentId,
-}: GetMemoComments) => {
-  try {
-    const result = await searchMemoComments.fetch({
-      memo_id: memoId,
-      limit: page ? LIMIT : undefined,
-      page,
-      depth,
-      parent: parentId,
-    });
-    return result;
-  } catch (err) {
-    toast.error(ERROR_ALERT);
-    return {
-      comments: [],
-      count: 0,
-      limit: 0,
-      page: 0,
-    };
+export const addNewComment = async (
+  req: PostReviewCommentReq | PostMemoCommentReq,
+  isReview: boolean
+) => {
+  let result;
+  if (isReview) {
+    result = await postReviewComment.fetch(req as PostReviewCommentReq);
+  } else {
+    result = await postMemoComment.fetch(req as PostMemoCommentReq);
   }
-};
-
-export const addReviewComment = async (req: PostReviewCommentReq) => {
-  const result = await postReviewComment.fetch(req);
-  return result;
-};
-
-export const addMemoComment = async (req: PostMemoCommentReq) => {
-  const result = await postMemoComment.fetch(req);
   return result;
 };
